Rename setters so they actually have "set" in their name
CanDeviceInterface's virtual methods are only used for setting the callback
functions which are actually called later on by the interface's implementation.
Meanwhile, their current names suggest as if they were the callback functions
themselves and their naming is similar to the one found in Qt's signals.

Since they are actually only setters for the said callbacks, add "set" to the
names of the functions to avoid confusion.

As a bonus, run clang-format on the affected files and move qRegisterMetaType in
candevicetest.cpp to main, since it can be registered for all the test cases
without causing any trouble.