Require the client to pass a CanDeviceInterface implementation to CanDevice.
Refactor the CanDevice classes by doing the following :

 * Throwing out CanFactory classes. Having an abstract base class for a factory
makes sense only if the actual implementing classes will use the same interface.
This wouldn't be possible with projected usage of this interface, where the next
factory implementation would create objects implementing a CAN-over-IP protocol,
which require vastly different constructor parameters compared to objects
created by QCanBus.

 * Requiring CanDevice's user to create a CanDeviceInterface object itself and
pass a reference to it via CanDevice's constructor. This allows the client to
choose the appropriate implementation by itself without having to rely on
factories. Additionally, it doesn't hard-code the parameters of CanDevice's
init() method to be the same as the factory's parameters. Furthermore, the
client can catch any errors coming from creating a CanDeviceInterface's derived
object itself, without the need of a factory proxying them. Last but not least,
a reference to the object is passed, which means that any null-checks inside
CanDevice become unnecessary.